Types: fix bug in TypeUtils.toString(Type) method
When string-izing a generic type which contains a type variable (or,
theoretically, a wildcard) whose bounds include itself, the method
would previously recurse infinitely, resulting in StackOverflowError.

The solution here is to only enumerate a type's bounds the first time it
is encountered; after that, we simply return the name of the type alone.